requirements.txt
numpy==1.19.5
scikit-image==0.17.2
torch==1.90
torchvision==0.10.0
pillow==8.1.1
opencv-python==4.5.3.56
paddlepaddle==2.4.1
paddlehub==2.3.1
gradio==3.40.1
